requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
fcm-django==0.3.10
idna==2.10
pyfcm==1.4.9
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
